edit for long values

+++ c_bst/lib/BST_actions.h
@@ -5,7 +5,7 @@ struct leaf {
     struct leaf *left;
     struct leaf *right;
 
-    int value;
+    long value;
 };
 
 struct leaf *create_leaf(void);
@@ -14,11 +14,11 @@ struct leaf *min_tree(struct leaf **tree);
 
 void transplant(struct leaf **rm_branch, struct leaf **branch);
 
-void delete(struct leaf **tree, int value);
+void delete(struct leaf **tree, long value);
 
-void insert(struct leaf **tree, int value);
+void insert(struct leaf **tree, long value);
 
-void search(struct leaf *tree, int value);
+void search(struct leaf *tree, long value);
 
 void dump_tree(struct leaf *tree);
